requirements.txt
alembic==1.7.7
click==8.1.2
colorama==0.4.4
cx-Oracle==8.3.0
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mysqlclient==2.1.0
packaging==21.3
pyparsing==3.0.8
python-dotenv==0.20.0
SQLAlchemy==1.4.35
uWSGI==2.0.20
webargs==8.1.0
Werkzeug==2.1.1
zipp==3.8.0
